requirements.txt
apache-airflow
pandas
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
tqdm